+++ src/filters/fb_network_builder.rs
@@ -92,6 +92,26 @@ impl<'a, 'f> FlatSerialize<'a, EngineFlatBuilder<'a>>
             FlatSerialize::serialize(o, builder)
         });
 
+        let opt_to_domains = network_filter.opt_to_domains.as_ref().map(|v| {
+            let mut o: Vec<u32> = v
+                .iter()
+                .map(|x| builder.get_or_insert_unique_domain_hash(x))
+                .collect();
+            o.sort_unstable();
+            o.dedup();
+            FlatSerialize::serialize(o, builder)
+        });
+
+        let opt_not_to_domains = network_filter.opt_not_to_domains.as_ref().map(|v| {
+            let mut o: Vec<u32> = v
+                .iter()
+                .map(|x| builder.get_or_insert_unique_domain_hash(x))
+                .collect();
+            o.sort_unstable();
+            o.dedup();
+            FlatSerialize::serialize(o, builder)
+        });
+
         let modifier_option = network_filter
             .modifier_option
             .map(|s| builder.create_string(s));
@@ -135,6 +155,8 @@ impl<'a, 'f> FlatSerialize<'a, EngineFlatBuilder<'a>>
                 modifier_option,
                 opt_domains,
                 opt_not_domains,
+                opt_to_domains,
+                opt_not_to_domains,
                 hostname,
                 tag,
                 raw_line,